Update text.html 10/8
I have started testing background colors and I learned that the border of a paragraph is a rectangle and I am still trying to find a way if I could turn it into a circle. A thought of mine is "Would I have to make a shape and to get that shape on the web?" Another curiosity I had was when playing with text fonts. I kept changing the fonts and the text started overlapping, I tried changing margin bottom, top and padding bottom and top and none of it worked. Tomorrow I will find a way to start the overlapping of the text.