Update test.html 10/14
Today I am worked with chatgpt to help make sure what I would start with if I was trying to create a Rock Paper Scissors game, and I learned that these first few lines help structure my website allowing it to be in english, having the right charset: makes it possible to use emojis and other characters that aren't in the traditional ASCII character set on the webpage. It also allows the right amount of spacing for the screen that the user has. I will begin to add CSS later.