Update test.html 11/7
I worked on trying to add a scroll feature, but it didn't work out. Instead I used padding and spacing to add products in rows and columns. This method seemed to be more efficient because I am not willing to learn how to scroll through images and manipulate their locations. I think today was a pretty effective day and tomorrow I will work on how to implement a login system.